Rename proxy package to destination package
