+++ controller/api/destination/endpoint_listener.go
@@ -1,4 +1,4 @@
-package proxy
+package destination
 
 import (
 	"fmt"
