+++ controller/api/destination/fallback_profile_listener_test.go
@@ -1,4 +1,4 @@
-package proxy
+package destination
 
 import (
 	"testing"
