+++ controller/api/destination/server.go
@@ -1,4 +1,4 @@
-package proxy
+package destination
 
 import (
 	"context"
