+++ controller/api/destination/test_helper.go
@@ -1,4 +1,4 @@
-package proxy
+package destination
 
 import (
 	"context"
